Introducing new laser module
